refactor(api): rename-move mass-create-user-accounts script
